Break up main file
Breaking main file into two smaller files
